Included jimple2cpg functionality and updated tests
