+++ src/main/scala/com/github/plume/oss/Jimple2Cpg.scala
@@ -2,13 +2,12 @@ package com.github.plume.oss
 
 import com.github.plume.oss.drivers.{IDriver, OverflowDbDriver}
 import com.github.plume.oss.passes._
-import com.github.plume.oss.passes.concurrent.{
-  PlumeCfgCreationPass,
-  PlumeContainsEdgePass,
-  PlumeHashPass
-}
-import com.github.plume.oss.passes.forkjoin.{PlumeCdgPass, PlumeCfgDominatorPass}
-import com.github.plume.oss.passes.parallel._
+import com.github.plume.oss.passes.base.{AstCreationPass, PlumeContainsEdgePass, PlumeCpgPassBase, PlumeFileCreationPass, PlumeMetaDataPass, PlumeMethodDecoratorPass, PlumeMethodStubCreator, PlumeNamespaceCreator, PlumeTypeDeclStubCreator, PlumeTypeNodePass}
+import com.github.plume.oss.passes.incremental.{PlumeDiffPass, PlumeHashPass}
+import com.github.plume.oss.passes.controlflow.PlumeCfgCreationPass
+import com.github.plume.oss.passes.controlflow.cfgdominator.PlumeCfgDominatorPass
+import com.github.plume.oss.passes.controlflow.codepencegraph.PlumeCdgPass
+import com.github.plume.oss.passes.reachingdef._
 import com.github.plume.oss.util.ProgramHandlingUtil
 import com.github.plume.oss.util.ProgramHandlingUtil.{extractSourceFilesFromArchive, moveClassFiles}
 import io.joern.x2cpg.SourceFiles
