+++ src/test/scala/com/github/plume/oss/querying/MethodParameterTests.scala
@@ -16,7 +16,7 @@ class MethodParameterTests extends Jimple2CpgFixture {
       """.stripMargin
 
   "should return exactly three parameters with correct fields" in {
-    cpg.parameter.filter(_.method.name == "foo").name.toSet shouldBe Set("this", "param1", "param2")
+    cpg.parameter.filter(_.method.name == "foo").name.toSetMutable shouldBe Set("this", "param1", "param2")
 
     val List(t) = cpg.parameter.filter(_.method.name == "foo").name("this").l
     t.code shouldBe "this"
